Improve video/audio synchronization.
Video and audio are better in sync now. This is achieved by offsetting video PTS by the timestamp of the first received Iframe and the audio PTS by the first received audio packet.

It's still not perfect and depends on what player is used (looks good in mpv and vlc, less so in Openshot).