Set default framerate to 30 FPS
This is the value most other camera apps ship as default.
The change will not affect existing users as default shared pref values are written to file on first access (meaning that the previous default of 60 is now saved in the pref files of anyone who has used the app before this change).

Closes #16